fix: Fixing status codes
