+++ internal/handlers/handlers.go
@@ -2,11 +2,9 @@ package handlers
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
-	"io/ioutil"
 	"net/http"
 	"sort"
 	"strconv"
@@ -48,94 +46,6 @@ func NewWithStorage(st *storage.Storage) *Handler {
 
 func (h *Handler) GetRouter() chi.Router {
 	return h.router
-}
-
-func (h *Handler) Update(w http.ResponseWriter, r *http.Request) {
-	ct := r.Header.Get("Content-Type")
-	if ct != "application/json" {
-		msg := "wrong content type, 'application/json' needed"
-		http.Error(w, msg, http.StatusUnsupportedMediaType)
-		return
-	}
-
-	respBody, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-	defer r.Body.Close()
-
-	m := metrics.Metrics{}
-	err = json.Unmarshal(respBody, &m)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusNotAcceptable)
-		return
-	}
-
-	switch m.MType {
-	case "gauge":
-		h.storage.PutGauge(m.ID, metrics.Gauge(*m.Value))
-	case "counter":
-		h.storage.PostCounter(m.ID, metrics.Counter(*m.Delta))
-	default:
-		err = fmt.Errorf("not implemented")
-		http.Error(w, err.Error(), http.StatusNotImplemented)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-}
-
-func (h *Handler) Value(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	ct := r.Header.Get("Content-Type")
-	if ct != "application/json" {
-		msg := "wrong content type, 'application/json' needed"
-		http.Error(w, msg, http.StatusUnsupportedMediaType)
-		return
-	}
-
-	reqBody, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	defer r.Body.Close()
-
-	m := metrics.Metrics{}
-	err = json.Unmarshal(reqBody, &m)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusNotAcceptable)
-		return
-	}
-
-	switch m.MType {
-	case "gauge":
-		gauge, errGet := h.storage.GetGauge(m.ID)
-		if errGet != nil {
-			http.Error(w, errGet.Error(), http.StatusNotFound)
-			return
-		}
-		val := float64(gauge)
-		m.Value = &val
-	case "counter":
-		h.storage.PostCounter(m.ID, metrics.Counter(*m.Delta))
-	default:
-		err = fmt.Errorf("not implemented")
-		http.Error(w, err.Error(), http.StatusNotImplemented)
-		return
-	}
-
-	body, err := json.Marshal(&m)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	w.Write(body)
 }
 
 func (h *Handler) Post(w http.ResponseWriter, r *http.Request) {
